requirements.txt
certifi==2021.5.30
chardet==3.0.4
idna==2.8
numpy==1.21.1
pandas==1.1.3
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.21.0
six==1.16.0
urllib3==1.24.3
